[UPDATE] EventService การอ่าน Ticket เข้าร่วมกิจกรรมจะเพิ่ม Participate ด้วย

+++ Microservice/UserService/src/main/java/meetu/userservice/model/InterestGenreBehavior.java
@@ -50,7 +50,7 @@ public class InterestGenreBehavior implements Comparable<InterestGenreBehavior> 
         return totalParticipate;
     }
 
-    public void setTotalParticipate(short totalParticipate) {
+    public void setTotalParticipate(int totalParticipate) {
         this.totalParticipate = totalParticipate;
     }
